Fix for signed/unsigned comparison with VectorField #rb none
#ROBOMERGE-OWNER: ryan.gerleve
#ROBOMERGE-AUTHOR: rob.krajcarski
#ROBOMERGE-SOURCE: CL 14176863 via CL 14176865 via CL 14176871 via CL 14176878 via CL 14176881
#ROBOMERGE-BOT: NETWORKING (Main -> Dev-Networking) (v723-14217278)

[CL 14230300 by rob krajcarski in Dev-Networking branch]

+++ Engine/Source/Runtime/Engine/Private/VectorField.cpp
@@ -416,7 +416,7 @@ void UVectorFieldStatic::UpdateCPUData(bool bDiscardData)
 			UE_LOG(LogVectorField, Warning, TEXT("SourceData.GetCopy() is supposed to unload the data after copying, but it is still loaded."));
 		}
 
-		const int32 SampleCount = SizeX * SizeY * SizeZ;
+		const size_t SampleCount = SizeX * SizeY * SizeZ;
 
 		FMemoryWriter Ar(CPUData);
